UIPlayerController updated
Added MouseNavigation, simply put the mouse uses a different input message so basically needs to do the same thing but not QUITE the same thing annoyingly.

Added a validation for clicking, so its sure that you submit while actually over that specific object that is to be submitted

animations are now triggered instead of a caret to kind of outline that this character has been selected

ValidateFlow updated so that it filters the types of selectables, these tend to do different things but act on submit

Added Enable/Disable input methods which simply sets hasSelected to true or false; this determines whether the input system messages actually run.