Natalyachanges
Added UI buttons to all 3 games.
Created menu screen with all buttons, the only thing cannot put background to the back - it seems to be on the front.
On the memory cards game slightly moved display to the left and now cards need to
be repositioned, i dont know how so left it to you.
Let me know if I have done it correctly, its my first time doing this appreciate your patience and understanding!